ft_printf no bonus completed

+++ ft_printf/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dground <dground@student.21-school.ru>     +#+  +:+       +#+        */
+/*   By: dground <dground@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 16:37:18 by dground           #+#    #+#             */
-/*   Updated: 2021/11/24 17:22:15 by dground          ###   ########.fr       */
+/*   Updated: 2021/11/24 19:22:21 by dground          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,6 @@ void	p_type(va_list arguments, int *length, char *str);
 int		ft_putchar(char c, int *length);
 void	ft_putnbr(int nb, int *length);
 void	ft_putuns(unsigned int nb, int *length);
-void	ft_puthex(unsigned int nb, int *length, char *str);
+void	ft_puthex(unsigned long long nb, int *length, char *str);
 
 #endif
